# Reuse buffer to reads gRPC messages  from the underlying reader