@adtac 2019-12-02
@alextomaili could you please run some benchmarks measuring the performance improvement? I'm particularly interested in the bytes allocated, QPS and latency stats. The reason I ask is, while the change may seem innocuous, I found a similar change to produce a lower memory footprint at the cost of a QPS/latency regression for a lot of common use-cases (e.g. small messages < 1 KiB in size) in my work to reuse Marshal byte buffers: https://github.com/grpc/grpc-go/pull/3167 - this is because of the overhead introduced by `sync.Pool`.

But a threshold similar to https://github.com/grpc/grpc-go/pull/3167 might not be straightforward because we don't know how big the message is going to be?