@adtac 2019-12-02
I suggest measuring these stats for messages of size 2^n for n in [0, 15]. See `benchmarks/benchmain`  for more details.