@alextomaili 2019-12-07
@adtac how about use "striped locks" pattern to reduce overhead of sync.Pool
we can have several sync.Pool instances instead one global instance and select pool by ring.

```
type (
	recvBufferPools struct {
		current             uint32
		bucketCount         uint32
		initialBufferSize   int
		maxPooledBufferSize int
		pools               []*sync.Pool
	}
)

var recvBufferPool = newRecvBufferPools(recvBufferPoolBucketsCount, initialRecvBufferSize, maxPooledRecvBufferSize)

func newRecvBufferPools(bucketCount uint32, initialBufferSize int, maxPooledBufferSize int) *recvBufferPools {
	p := &recvBufferPools{
		bucketCount:         bucketCount,
		initialBufferSize:   initialBufferSize,
		maxPooledBufferSize: maxPooledBufferSize,
		pools:               make([]*sync.Pool, 0, bucketCount),
	}

	var i uint32
	for i = 0; i < p.bucketCount; i++ {
		p.pools = append(p.pools, &sync.Pool{
			New: func() interface{} {
				return &recvBuffer{
					buf: make([]byte, p.initialBufferSize),
				}
			},
		})
	}

	return p
}

func (p *recvBufferPools) getBuffer() *recvBuffer {
	i := atomic.AddUint32(&p.current, 1) % p.bucketCount
	b := p.pools[i].Get().(*recvBuffer)
	b.poolIndex = i
	return b
}

func (p *recvBufferPools) freeBuffer(b *recvBuffer) {
	if b != nil && b.cap() < p.maxPooledBufferSize && b.poolIndex < p.bucketCount {
		p.pools[b.poolIndex].Put(b)
	}
}
```

This way can increase probability to take pooled object by fast path, also it can reduce contention on mutex inside sync.Pool if fast  path will be failed. Of course this way also increases probability to call memory allocation when selected pool is empty. 

I've try used 32 instance of sync.Pool and got positive results for some cases

for example 

**maxConcurrentCalls_6-reqSize_8192B-respSize_8192B - one sync.Pool**
```
unary-networkMode_none-bufConn_false-keepalive_false-benchTime_40s-trace_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_6-reqSize_8192B-respSize_8192B-compressor_off-channelz_false-preloader_false
               Title       Before        After Percentage
            TotalOps       229956       214257    -6.83%
             SendOps            0            0      NaN%
             RecvOps            0            0      NaN%
            Bytes/op    107606.77     90527.60   -15.87%
           Allocs/op       187.89       186.29    -0.53%
             ReqT/op 376759910.40 351038668.80    -6.83%
            RespT/op 376759910.40 351038668.80    -6.83%
            50th-Lat    475.679µs    459.293µs    -3.44%
            90th-Lat    912.167µs    922.482µs     1.13%
            99th-Lat   1.699435ms   1.868889ms     9.97%
             Avg-Lat   1.042933ms   1.119393ms     7.33%
```

**maxConcurrentCalls_6-reqSize_8192B-respSize_8192B - 32 instance of sync.Pool**

```
unary-networkMode_none-bufConn_false-keepalive_false-benchTime_40s-trace_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_6-reqSize_8192B-respSize_8192B-compressor_off-channelz_false-preloader_false
               Title       Before        After Percentage
            TotalOps       229956       370150    60.97%
             SendOps            0            0      NaN%
             RecvOps            0            0      NaN%
            Bytes/op    107606.77    113474.69     5.45%
           Allocs/op       187.89       191.77     2.13%
             ReqT/op 376759910.40 606453760.00    60.97%
            RespT/op 376759910.40 606453760.00    60.97%
            50th-Lat    475.679µs    455.601µs    -4.22%
            90th-Lat    912.167µs    967.353µs     6.05%
            99th-Lat   1.699435ms   2.158552ms    27.02%
             Avg-Lat   1.042933ms    647.638µs   -37.90%
```

significant profit here

but for some cases i've got small degradation

**maxConcurrentCalls_6-reqSize_1024B-respSize_1024B - one sync.Pool**

```
unary-networkMode_none-bufConn_false-keepalive_false-benchTime_40s-trace_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_6-reqSize_1024B-respSize_1024B-compressor_off-channelz_false-preloader_false
               Title       Before        After Percentage
            TotalOps       734872       658622   -10.38%
             SendOps            0            0      NaN%
             RecvOps            0            0      NaN%
            Bytes/op     20943.92     18722.08   -10.60%
           Allocs/op       182.38       180.19    -1.10%
             ReqT/op 150501785.60 134885785.60   -10.38%
            RespT/op 150501785.60 134885785.60   -10.38%
            50th-Lat    295.356µs    314.253µs     6.40%
            90th-Lat    433.068µs    526.701µs    21.62%
            99th-Lat    834.935µs   1.088915ms    30.42%
             Avg-Lat    325.959µs     363.72µs    11.58%
```

**maxConcurrentCalls_6-reqSize_1024B-respSize_1024B - 32 sync.Pool**

```
unary-networkMode_none-bufConn_false-keepalive_false-benchTime_40s-trace_false-latency_0s-kbps_0-MTU_0-maxConcurrentCalls_6-reqSize_1024B-respSize_1024B-compressor_off-channelz_false-preloader_false
               Title       Before        After Percentage
            TotalOps       734872       652726   -11.18%
             SendOps            0            0      NaN%
             RecvOps            0            0      NaN%
            Bytes/op     20943.92     21520.63     2.75%
           Allocs/op       182.38       182.10     0.00%
             ReqT/op 150501785.60 133678284.80   -11.18%
            RespT/op 150501785.60 133678284.80   -11.18%
            50th-Lat    295.356µs    310.461µs     5.11%
            90th-Lat    433.068µs    519.426µs    19.94%
            99th-Lat    834.935µs   1.131208ms    35.48%
             Avg-Lat    325.959µs    366.967µs    12.58%
```

small degradation here

I think this is not bad idea if we can't use sync.Pool as attribute of some high level abstraction such as "connection", "session" etc by easy way. (each "connection" or "session" can use own pool and we will have less contentions on pool rather than we have one global pool)

If we want to use pool deep inside low level abstraction such as "parser" or "codec" we can use ring of several pools do reduce overhead.

What do you think about this ?